Extract release hooks into separated modules
