+++ cli/definitions/cli.ts
@@ -1,4 +1,7 @@
-﻿export enum CommandType {
+﻿/// <reference path="../../definitions/generated/code-push.d.ts" />
+import AccountManager = require("code-push");
+
+export enum CommandType {
     accessKeyAdd,
     accessKeyPatch,
     accessKeyList,
@@ -220,4 +223,9 @@ export interface ISessionRemoveCommand extends ICommand {
     machineName: string;
 }
 
-export type ReleaseHook = (currentCommand: IReleaseCommand, originalCommand: IReleaseCommand) => Q.Promise<IReleaseCommand|void>;
+export type ReleaseHook = (currentCommand: IReleaseCommand, originalCommand?: IReleaseCommand, sdk?: AccountManager) => Q.Promise<IReleaseCommand|void>;
+
+export interface ReleaseFile {
+    sourceLocation: string;     // The current location of the file on disk
+    targetLocation: string;     // The desired location of the file within the zip
+}
