+++ cli/script/release-hooks/index.ts
@@ -1,0 +1,6 @@
+import {ReleaseHook} from "../../definitions/cli";
+
+export var hooks: ReleaseHook[] = [
+  require('./signing'),
+  require('./core-release'),
+];
